# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Roayti - AI-powered novel writing and reading platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /library/{novelId}:
    post:
      operationId: addToLibrary
      tags: [library]
      summary: Add novel to library
      parameters:
        - name: novelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryItem"
    delete:
      operationId: removeFromLibrary
      tags: [library]
      summary: Remove novel from library
      parameters:
        - name: novelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /library/{novelId}/progress:
    put:
      operationId: updateReadingProgress
      tags: [library]
      summary: Update reading progress
      parameters:
        - name: novelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgressBody"
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryItem"
...
  /gemini/conversations:
    get:
      operationId: listGeminiConversations
      tags: [gemini]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiConversation"
    post:
      operationId: createGeminiConversation
      tags: [gemini]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGeminiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversation"
  /gemini/conversations/{id}:
    get:
      operationId: getGeminiConversation
      tags: [gemini]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
    delete:
      operationId: deleteGeminiConversation
      tags: [gemini]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
  /gemini/conversations/{id}/messages:
    get:
      operationId: listGeminiMessages
      tags: [gemini]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiMessage"
    post:
      operationId: sendGeminiMessage
      tags: [gemini]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendGeminiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
  /gemini/generate-image:
    post:
      operationId: generateGeminiImage
      tags: [gemini]
      summary: Generate an image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateGeminiImageBody"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateGeminiImageResponse"
components:
  schemas:
...
    Novel:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
          nullable: true
        genre:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published]
        authorId:
          type: integer
        authorClerkId:
          type: string
        authorName:
          type: string
        coverImage:
          type: string
          nullable: true
        likesCount:
          type: integer
        viewsCount:
          type: integer
        chaptersCount:
          type: integer
        isLiked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - authorId
        - authorClerkId
        - authorName
        - likesCount
        - viewsCount
        - chaptersCount
        - isLiked
        - createdAt
        - updatedAt
...
    LibraryItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        novelId:
          type: integer
        novel:
          $ref: "#/components/schemas/Novel"
        lastReadChapterId:
          type: integer
          nullable: true
        progress:
          type: integer
        addedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - novelId
        - novel
        - progress
        - addedAt
    UpdateProgressBody:
      type: object
      properties:
        lastReadChapterId:
          type: integer
        progress:
          type: integer
      required:
        - progress
...
    GeminiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    GeminiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateGeminiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendGeminiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    GeminiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/GeminiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    GenerateGeminiImageBody:
      type: object
      properties:
        prompt:
          type: string
      required:
        - prompt
    GenerateGeminiImageResponse:
      type: object
      properties:
        b64_json:
          type: string
        mimeType:
          type: string
      required:
        - b64_json
        - mimeType
    GeminiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error